Update README.md
It has minimal design with HTML and CSS.
A text box at the top to add a todo item to the list.
Add button: when a user clicks on the button, add that todo item to the list. Also we use 'enter' key to save task into list.
Using localStorage to store the user data, when they goes off again able to retrive their old non deleted tasks.
A delete button to delete the todo item completely from the list.
A checkbox to check the item off the list. (checked item in the list, with some differentiation from rest of the todo items)
Total number of tasks in the list.